Dark Jesus Collab Adding Mobs and Buffing in Tet
* Dark Jesus Collab Adding Mobs and Buffing

Drop Creepers can now spawn any time of day. Truely, I've never seen these mobs while I was in Concordia, so hopefully we see some now.

Jungle skelly added. Has similar powers to the Fire Watch Skelly with the knockback and all, but a third of the health and a larger range for it to spawn (only one should spawn within 80 blocks of a player).

Squidder joke mob for squids added, too, with 4% chance

Firespider mob added to buff difficulty of mining obby on the obby islands. 40% chance of spawning, and has a nasty wither bite.

Also, added zombie lumberjacks to the empty, decaying town of BongoBongo. May it rest in peace and hopefully the zombie loggers keep the travelers from tearing down the last of Tet's jungle. (45% chance of spawning per 2 minutes within 32 meters of player.) Also, he drops a fire aspect wood axe and has a speed 1 buff.

* Spawner and glass helmet deleted