DJ Collab: Lots of Proposed Spawn Changes to Zion
* DJ Collab: Lots of Proposed Spawn Changes to Zion

To keep with the lore of how hellish the Hell biome and surrounding areas is.

DarkJesus implemented fire rain. We hope it works once per hour (or more in the future). The fireballs spawn within 96 meters of the player, too. The fireballs should not destroy terrain or set new fires, too.

Mob spawning on the extreme hills has been buffed. Due to how the config once had a 1 minute spawn time instead of 15 seconds like other shards, it is no wonder why we never saw mobs in the extreme hills! Mobs were far too unlikely to spawn. Also, due to Zion's limited capacity, mobs are set to despawn upon the player unloading the chunk. 

We also asked for custom mobs on the subreddit. Only the Spirit of Sintralin made the list, and it makes sense that a banned player would be nothing more than a zombie in Zion's Hell.

Also, we changed some spawn configs of your White Walker zombie. It feels obvious that it was not spawning because they needed to spawn in snow AND in air. We'll see if they spawn this time.

Also, added an amored wither skelly that drops only a fire aspect bone. Possibly going to make it harder. 

Hopefully, the hell biome is more challenging as it should be for being hell. More buffing may come.

* Fixed fireball spawning, edited Sintralin out

The fireball will now spawn every minute and will impact within 16 meters of a player in the Hell biome.

The Sintralin mob is now a Succbus. The renaming should keep people from asking questions about its name.