DarkJesus Collab to Further Buff Fire Watch in Aleph
* DarkJesus Collab to Further Buff Fire Watch

We replaced the diamond chestplate with leather, and the skelly is all leather armor now to reflect the lore of an ancient forest watch that is annoyed with birch forest destruction. The skelly now carries a powerful blaze rod that has knockback 8, sharpness 2. In addition, a hit by this item now induces blindness and wither effects.

Also, it's important to note that this skelly is wearing a black glass head. We are not sure if that line of code to add a glass head will work, but if it does not, then get rid of the glass head. If it works, then we have skellies with black glass heads.

* Boots will have one enchantment, only

Code can't handle two enchants

* Glass helmet eliminated, Feather boots returned

Mobs can't wear foreign blocks like black glass, so the code is changed to let the skelly wear the fire prot leather for a helmet. Also, the boots have feather falling again to make this thing harder to die from falling.

Further, a small dropchance of 2% remains on all items that this skelly wears, but the special blaze rod will not drop. (Eliminating that itemheld dropchance option should still make it impossible for the skelly to drop that blaze rod.) Those armor sets could be a novel item to walk around in, and should take a lot of time to grind for. 

Also, the skelly should not be able to pick up items anyway, so I got rid of the option that set Pickupitems to false since it is redundant.