DJ Collab to add Iron Golem and special witches to Bet
* DJ Collab to add Iron Golem and special witches to Bet

Like the previous pull, formatting of the name of the mobs may be interesting for testing the code. Also, the iron golem in this code only has a chance of dropping a fire aspect II sword, and it should not drop any iron. 

Also, there is a witch spawn set over the empty and ruined NSR. It's just a cobble encased town with grief all inside after a great war runined all of its newfriends. They are invisible, slow, and drop newfriend souls (netherwart lore) when they die. This should test if it's possible for witches to spawn in an area defined by a circle.

As for the iron golem, it is tough, and it may be set to create special, valuable ingredients in the future that require many players to kill it. I reccomend not hitting it if you want to live.

* Spawner Cut out, Indenting fixed

The attempt to create an invisible witch that spawns in the NSR failed. Until there is a way to make a mob spawn within a radius of this city, we'll leave that code out.

Also, indenting fixed on Iceplains. Good catch, Max.

* OP iron golem to drop 2 sticky pistons

People really wanted this thing to drop sticky pistons, or else it wasn't worth the trouble to kill.

* Despawn on chunk unload set to true on iron golem

DJ noticed suggested this, and he said it's a better idea to set despawning to true on chunk unload to make planning more difficult.Also, setting this to true ensures that people can't fill their vaults with the golem.